Refactor: Added single responsability on store user in redux action
