+++ src/redux/user/userActions.js
@@ -1,18 +1,7 @@
-import axios from 'axios';
+import { setAuthToken } from '@/config/axios';
 
 export function setUser(user) {
-  const storage = JSON.parse(localStorage.getItem('user'));
-
-  const currentToken = storage ? storage.token : null;
-  const newToken = user.token;
-  const token = newToken || currentToken;
-
-  if (token) {
-    axios.defaults.headers.common.Authorization = `bearer ${token}`;
-  } else {
-    delete axios.defaults.headers.common.Authorization;
-  }
-
+  setAuthToken(user.token);
   return {
     type: 'SAVE_USER',
     user: user.user,
